wip: moving all controllers to its own use cases separation
