+++ TodoApp/WebApi/Controllers/Todo/v1/TodoJsonControllerBase.cs
@@ -1,0 +1,10 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace TodoApp.WebApi.Controllers.Todo.v1
+{
+    [ApiController]
+    [Route("v1/todo")]
+    [Produces("application/json")]
+    public abstract class TodoJsonControllerBase : ControllerBase
+    { }
+}
